+++ src/components/App.scss
@@ -1,0 +1,6 @@
+.App {
+  text-align: center;
+  width: 100vw;
+  height: 100vh;
+  background: linear-gradient(to right, #1578ec, #65cafa);
+}
