+++ src/components/App.tsx
@@ -1,0 +1,12 @@
+import "./App.scss";
+import Logo from "./Logo";
+
+function App() {
+  return (
+    <div className="App">
+      <Logo />
+    </div>
+  );
+}
+
+export default App;
