+++ src/components/Logo.scss
