+++ src/main.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import Logo from "./Logo";
+import App from "./components/App";
 import "./index.scss";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <Logo />
+    <App />
   </React.StrictMode>
 );
